RABSW-497: Rebase dws-operator on operator-sdk 1.10
This commit starts with a fresh operator-sdk workspace generated by operator-sdk 1.10
and copies the reconcile logic from the existing dws-operator. For now, I haven't
copied any of the existing helm or Jenkins infrastructure.

Operator-sdk 1.10 added support for making webhooks. In this commit I also generated
webhooks for validating and mutating the workflow resource. I copied the logic from the
dws-admission-webhook repo with modifications to make it work in the new layout. This
involved:
 - Better splitting apart of the validating vs. mutating portions of the webhook
 - Changing/adding environment variables. Since these webhooks are part of the controller
   docker image, the environment variables needed to have more specific names
 - Making it possible to run the mutating webhook as part of the nnf-sos container

With this new layout, a single validating webhook is run in the same pod as the dws-operator.
This webhook has a list of rulesets that it validates against. The mutating webhooks live in
the pod with their driver (e.g., nnf-sos) and only handle mutating the workflow object.