RABSW-508: Add Jenkins and helm
This commit adds back Jenkins and helm. This is not using the cray-service chart.

There are now two ways to deploy in a kind environment. You can use either:

make deploy

or

helm install kubernetes/dws-operator

These use two different sets of yaml files since the two methods (kustomize and helm)
are too different to share configs. I had to make some more changes to the way the webhook
finds the rulesets because the namespace we were looking in was hardcoded to assume we were
using kustomize.