RABSW-192: Mount support for physically connected storage (#95)
This commit changes the client mount code to allow mounting PCIe connected LVM
storage. The compute node should be able to see the logical volumes created on
the Rabbit nodes, but they need to be activated before they can be used.

Also, the ClientMount resource was changed to allow specifying an object reference
to an NnfNodeStorage resource for mounting file systems on the Rabbit. The clientmountd
daemon is not used on the Rabbit node.

Co-authored-by: Matt Richerson <mattr@cray.com>