RABSW-1007: Check for unique fields (#38)
* RABSW-1007: Check for unique fields

This commit adds a new option for a rule definition, "uniqueWithin". This
option allows rules to specify that a key/value pair needs to be unique across
multiple #DWs. An example is the "name" field in a jobdw directive. Multiple
jobdws should not use the same value for name. However, The same value for name
can be used between a jobdw and a create_persistent. "uniqueBetween" allows for
this by specifying a key. Values will only be checked for uniqueness for rule
definitions using the same "uniqueBetween" key.

Signed-off-by: Matt Richerson <mattr@cray.com>

* formatting change

Signed-off-by: Matt Richerson <mattr@cray.com>

Signed-off-by: Matt Richerson <mattr@cray.com>
Co-authored-by: Matt Richerson <mattr@cray.com>