Change mount daemon name to clientmountd
The Makefile target build-daemon builds an executable called
`clientmountd` and the specfile also installs it to
/usr/bin/clientmountd. When doing an install of the daemon, the
service file created is called `clientmount.service` rather than
`clientmountd.service`.

Changing the name here allows for consistency, which is necessary for
nnf-deploy to use the dws Makefile to build the daemon rather than
building itself. With the change to using the Makefile, nnf-deploy will
then look for a service named `clientmountd` rather than `clientmount`.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>